--- client/app/src/main/java/com/example/criengine/1029.MainActivity.java
@@ -1,117 +1,0 @@
-package com.example.criengine;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-
-import com.firebase.ui.auth.AuthUI;
-import com.firebase.ui.auth.IdpResponse;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.firestore.DocumentSnapshot;
-
-import java.util.Arrays;
-import java.util.List;
-
-public class MainActivity extends AppCompatActivity {
-    private Button myProfileActivityButton;
-    private Button myBooksActivityButton;
-    private Button requestActivityButton;
-    private Button myBookActivityButton;
-    DatabaseWrapper dbw; // needs to be initialized by sign in
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-
-        requestActivityButton = findViewById(R.id.requested_books_activity);
-
-        requestActivityButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(v.getContext(), RequestActivity.class);
-                v.getContext().startActivity(intent);
-            }
-        });
-
-        // router to other activities for testing
-        myProfileActivityButton = findViewById(R.id.my_profile_activity_button);
-
-        myProfileActivityButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(v.getContext(), MyProfileActivity.class);
-                v.getContext().startActivity(intent);
-            }
-        });
-
-        myBooksActivityButton = findViewById(R.id.myBooksButton);
-
-        myBooksActivityButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(v.getContext(), MyBooksActivity.class);
-                v.getContext().startActivity(intent);
-            }
-        });
-
-            
-        // router to book activity
-        myBookActivityButton = findViewById(R.id.my_book_activity_button);
-
-        myBookActivityButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(v.getContext(), MyBookActivity.class);
-                v.getContext().startActivity(intent);
-            }
-        });
-
-        //begin sign in
-        // accounts: user1@email.com:password
-        //           user2@email.com:password
-        List<AuthUI.IdpConfig> providers = Arrays.asList(
-                new AuthUI.IdpConfig.EmailBuilder().build());
-        startActivityForResult(
-                AuthUI.getInstance()
-                        .createSignInIntentBuilder()
-                        .setAvailableProviders(providers)
-                        .setIsSmartLockEnabled(false)
-//                        .setLogo(R.drawable.my_great_logo)
-//                        .setTheme(R.style.MySuperAppTheme)
-                        .build(),
-                1);
-
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-
-        if (requestCode == 1) {
-            IdpResponse response = IdpResponse.fromResultIntent(data);
-
-            if (resultCode == RESULT_OK) {
-                // Successfully signed in
-                FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
-                System.out.println(user);
-//                dbw.setUser(user);
-                // ...user1
-            } else {
-                // Sign in failed. If response is null the user canceled the
-                // sign-in flow using the back button. Otherwise check
-                // response.getError().getErrorCode() and handle the error.
-                // ...
-            }
-        }
-    }
-}
